backend/requirements.txt
fastapi
uvicorn
tortoise-orm
passlib[bcrypt]
python-jose
bcrypt